+++ projects/pastanaga-angular/src/lib/avatar/avatar.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AvatarComponent } from './avatar.component';
-import { PaIconModule } from '../icon/icon.module';
 
 @NgModule({
     declarations: [AvatarComponent],
     exports: [AvatarComponent],
-    imports: [CommonModule, PaIconModule],
+    imports: [CommonModule],
 })
 export class PaAvatarModule {}